+++ src/components/Table/TD/index.js
@@ -6,10 +6,15 @@ const TD = (props) => {
     'oui-numerical': props.isNumerical,
   });
 
+  const styles = {
+    width: props.width,
+  };
+
   return (
     <td
       className={ classes }
-      data-test-section={ props.testSection }>
+      data-test-section={ props.testSection }
+      style={ styles }>
       { props.children }
     </td>
   );
@@ -22,6 +27,8 @@ TD.propTypes = {
   isNumerical: React.PropTypes.bool,
   /** Hook for automated JavaScript tests */
   testSection: React.PropTypes.string,
+  /** A number with a unit that becomes the width of the `Table` cell */
+  width: React.PropTypes.string,
 };
 
 TD.defaultProps = {