+++ src/components/Table/TH/index.js
@@ -7,10 +7,15 @@ const TH = (props) => {
     'oui-cell-collapse': props.isCollapsed,
   });
 
+  const styles = {
+    width: props.width,
+  };
+
   return (
     <th
       className={ classes }
-      data-test-section={ props.testSection }>
+      data-test-section={ props.testSection }
+      style={ styles }>
       { props.children }
     </th>
   );
@@ -25,6 +30,8 @@ TH.propTypes = {
   isNumerical: React.PropTypes.bool,
   /** Hook for automated JavaScript tests */
   testSection: React.PropTypes.string,
+  /** A number with a unit that becomes the width of the `Table` cell */
+  width: React.PropTypes.string,
 };
 
 TH.defaultProps = {